+++ src/tools/DeepAnalysis/FnVar/def/OutputVarCommandBase.ts
@@ -1,10 +1,10 @@
 import type { TValMetaIn } from '../../../../AhkSymbol/CAhkFunc';
-import { replacerSpace } from '../../../str/removeSpecialChar';
+import { FindExprDelim } from '../../../zFromCpp/FindExprDelim';
 import type { TGetFnDefNeed } from '../TFnVarDef';
 import { getValMeta } from './getValMeta';
 /* cSpell:disable */
 
-const OutputVarCommandMap: ReadonlyMap<string, RegExp> = new Map([
+const OutputVarCommandMap: ReadonlyMap<string, number> = new Map([
     // usually case exp: EnvGet, v, %EnvVarName%
     'CATCH',
     'CONTROLGET',
@@ -50,44 +50,54 @@ const OutputVarCommandMap: ReadonlyMap<string, RegExp> = new Map([
     'WINGETCLASS',
     'WINGETTEXT',
     'WINGETTITLE',
-    // eslint-disable-next-line security/detect-non-literal-regexp
-].map((wordUp: string): [string, RegExp] => [wordUp, new RegExp(`(^\\s*${wordUp}[,\\s]*)`, 'iu')]));
+].map((wordUp: string): [string, number] => [wordUp, wordUp.length]));
+
+type TScanData = {
+    RawNameNew: string;
+    lPos: number;
+};
+
+function spiltCommandOne(lStr: string, quickMake: number): TScanData {
+    let make = quickMake;
+
+    const oldMake: number = 1 + FindExprDelim(lStr, ',', make + 1);
+    make = FindExprDelim(lStr, ',', oldMake);
+    const partStr: string = lStr.slice(oldMake, make);
+    const RawNameNew: string = partStr.trim();
+
+    return {
+        RawNameNew,
+        lPos: oldMake + partStr.indexOf(RawNameNew),
+    };
+}
 
 /**
  * OutputVar
  * [Other Functions](https://www.autohotkey.com/docs/Functions.htm#Other_Functions)
  * [Polyethene's Command Functions](https://github.com/polyethene/AutoHotkey-Scripts/blob/master/Functions.ahk):
  * Provides a callable function for each AutoHotkey command that has an OutputVar.
+ *
+ * 15.11ms 38.29ms
  */
 export function OutputVarCommandBase(arg: TGetFnDefNeed, fistWordUp: string): null {
-    const regexp: RegExp | undefined = OutputVarCommandMap.get(fistWordUp);
-    if (regexp === undefined) return null;
+    const quickMake: number | undefined = OutputVarCommandMap.get(fistWordUp);
+    if (quickMake === undefined) return null;
     //
-
     const {
         lStr,
-        valMap,
         line,
         paramMap,
         GValMap,
+        valMap,
         lineComment,
     } = arg;
 
-    // TODO remove ...regexp
-    const lStrFix = lStr.replace(regexp, replacerSpace);
+    const { RawNameNew, lPos } = spiltCommandOne(lStr, quickMake);
 
-    const v: RegExpMatchArray | null = lStrFix.match(/\b(\w+)\b/ui);
-    if (v === null) return null;
-
-    const character: number | undefined = v.index;
-    if (character === undefined) return null;
-
-    const RawName: string = v[1];
-    const UpName: string = RawName.toUpperCase();
+    const UpName: string = RawNameNew.toUpperCase();
     if (paramMap.has(UpName) || GValMap.has(UpName)) return null;
 
-    const value: TValMetaIn = getValMeta(line, character, RawName, valMap, lineComment);
+    const value: TValMetaIn = getValMeta(line, lPos, RawNameNew, valMap, lineComment);
     valMap.set(UpName, value);
-
     return null;
 }